Upgraded to doubles.

+++ controller.cpp
@@ -8,8 +8,8 @@ void Controller::init(){
     universe = new Universe(*renderer);
 
     Matter a (
-        Vec3<float> (50, 50, 0),
-        Vec3<float> (0),
+        Vec3<double> (50, 50, 0),
+        Vec3<double> (0),
         0
     );
     Renderer::Rectangle *gr_rect = new Renderer::Rectangle(20, 40); // This is a memory leak
@@ -17,8 +17,8 @@ void Controller::init(){
     universe->add_matter(a);
 
     Matter b (
-        Vec3<float> (300, 300, 0),
-        Vec3<float> (0),
+        Vec3<double> (300, 300, 0),
+        Vec3<double> (0),
         0
     );
     Renderer::Circle *gr_circ = new Renderer::Circle(20);   // This is a memory leak
